requirements.txt
flask>=3.0
requests>=2.32
twilio>=9.0

google-cloud-storage>=2.18
google-genai>=1.0

duckdb>=1.1
pandas>=2.2
numpy==1.26.4

opencv-python-headless>=4.10
paddlepaddle==3.2.2
paddleocr==3.5.0
paddlex==3.5.0
Pillow>=10.4

pydantic>=2.8,<3
python-dateutil>=2.9
python-dotenv>=1.0

transformers>=4.45
scikit-learn>=1.5
rapidfuzz>=3.9

groq>=0.11
anthropic>=0.34

streamlit>=1.35
plotly>=5.20
tabulate>=0.9
